# Imtihon 2/api_gataway/docs/swagger.yaml
definitions:
  genproto.CreateEducationRequest:
    properties:
      degree:
        type: string
      end_date:
        type: string
      field_of_study:
        type: string
      id:
        type: string
      institution:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  genproto.CreateExperienceRequest:
    properties:
      company:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.Education:
    properties:
      degree:
        type: string
      end_date:
        type: string
      field_of_study:
        type: string
      id:
        type: string
      institution:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.UpdateEducationRequest:
    properties:
      degree:
        type: string
      deleted_at:
        type: integer
      end_date:
        type: string
      field_of_study:
        type: string
      id:
        type: string
      institution:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  genproto.UpdateExperienceRequest:
    properties:
      company:
        type: string
      deleted_at:
        type: integer
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  title: Portfolio swagger UI
paths:
  /education/create:
    post:
      consumes:
      - application/json
      description: This API for creating education
      parameters:
      - description: Education
        in: body
        name: education
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateEducationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Create Education
      tags:
      - education
  /education/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting education
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Delete Education
      tags:
      - education
  /education/get/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting education
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Education'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Education
      tags:
      - education
  /education/getall:
    get:
      consumes:
      - application/json
      description: This API for getting educations
      parameters:
      - description: UserId
        in: query
        name: user_id
        type: string
      - description: Institution
        in: query
        name: institution
        type: string
      - description: Degree
        in: query
        name: degree
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Education'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Educations
      tags:
      - education
  /education/getallbyuser:
    get:
      consumes:
      - application/json
      description: This API for getting educations
      parameters:
      - description: UserId
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Education'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Educations By UserId
      tags:
      - education
  /education/update:
    put:
      consumes:
      - application/json
      description: This API for updating education
      parameters:
      - description: Education
        in: body
        name: education
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateEducationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Update Education
      tags:
      - education
  /experience/create:
    post:
      consumes:
      - application/json
      description: This API for creating experience
      parameters:
      - description: Experience
        in: body
        name: experience
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateExperienceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Create Experience
      tags:
      - experience
  /experience/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting experience
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Delete Experience
      tags:
      - experience
  /experience/get/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting experience
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Experience
      tags:
      - experience
  /experience/getall:
    get:
      consumes:
      - application/json
      description: This API for getting experiences
      parameters:
      - description: Experience User ID
        in: query
        name: user_id
        type: string
      - description: Company Name
        in: query
        name: company
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Experiences
      tags:
      - experience
  /experience/getallbyuser:
    get:
      consumes:
      - application/json
      description: This API for getting experiences
      parameters:
      - description: Experience User ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Experiences By User
      tags:
      - experience
  /experience/update:
    put:
      consumes:
      - application/json
      description: This API for updating experience
      parameters:
      - description: Experience
        in: body
        name: experience
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateExperienceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Update Experience
      tags:
      - experience
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"